backend/requirements/requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
psycopg2==2.9.9
channels==4.0.0
daphne==4.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0